Correct implementation of stable topological sort of dependency graphs
This fixes the dependency sorting problems that could occur with complex
dependency graphs. The previous implementation was overly complex and
on top of that, not actually correct. The new implementation is a very simple
topological sort of the graph, but using a priority queue based on
embeddings so that embeddings are always grouped together when
dependencies allow it.